Minor rework
